Rename ComposableAppTable to ComposableTableWithControls
