+++ client/src/app/pages/waves/waves.tsx
@@ -12,7 +12,7 @@ import {
 } from "@patternfly/react-core";
 import { useTranslation } from "react-i18next";
 import { useFetchWaves } from "@app/queries/waves";
-import { ComposableAppTable } from "@app/shared/components/composable-app-table";
+import { ComposableTableWithControls } from "@app/shared/components/composable-table-with-controls";
 import {
   AppPlaceholder,
   ConditionalRender,
@@ -31,7 +31,7 @@ import { useFilterState } from "@app/shared/hooks/useFilterState";
 import { useSelectionState } from "@migtools/lib-ui";
 import { TdProps } from "@patternfly/react-table";
 import dayjs from "dayjs";
-import { IComposableRow } from "@app/shared/components/composable-app-table/composable-app-table";
+import { IComposableRow } from "@app/shared/components/composable-table-with-controls/composable-table-with-controls";
 import { WaveApplicationsTable } from "./wave-applications-table/wave-applications-table";
 import { WaveStakeholdersTable } from "./wave-stakeholders-table/wave-stakeholders-table";
 import { CreateEditWaveModal } from "./components/create-edit-wave-modal";
@@ -278,7 +278,7 @@ export const Waves: React.FC = () => {
           when={isFetching && !(waves || fetchError)}
           then={<AppPlaceholder />}
         >
-          <ComposableAppTable
+          <ComposableTableWithControls
             isSelectable={true}
             handleToggleRowSelected={handleToggleRowSelected}
             handleIsRowSelected={handleIsRowSelected}
@@ -336,7 +336,7 @@ export const Waves: React.FC = () => {
               />
             }
             toolbarClearAllFilters={handleOnClearAllFilters}
-          ></ComposableAppTable>
+          ></ComposableTableWithControls>
         </ConditionalRender>
       </PageSection>
       <CreateEditWaveModal
