+++ client/src/app/shared/components/composable-table-with-controls/index.ts
@@ -1,0 +1,1 @@
+export { ComposableTableWithControls } from "./composable-table-with-controls";
